Split version and patch stores out
